@@ -8,10 +8,11 @@ wget https://github.com/thewh1teagle/kokoro-onnx/releases/download/model-files-v
 python examples/with_stream.py
 """
 
-import sounddevice as sd
-from kokoro_onnx import Kokoro
 import asyncio
 
+import sounddevice as sd
+
+from kokoro_onnx import Kokoro
 
 text = """
 We've just been hearing from Matthew Cappucci, a senior meteorologist at the weather app MyRadar, who says Kansas City is seeing its heaviest snow in 32 years - with more than a foot (30 to 40cm) having come down so far.
